requirements.txt
# Web Framework
Flask
Werkzeug

# Data Processing
# Pandas bản mới mới hỗ trợ Python 3.13
pandas>=2.2.3
# Numpy bản mới
numpy>=1.26.4
openpyxl
pyreadstat

# Data Visualization
matplotlib
seaborn

# Statistical Analysis
scipy
statsmodels

# Insight & Assistant Engine
groq

# Environment Variables
python-dotenv

# Production Server
gunicorn

python-docx